Feat: getContent 분리
